[arcane,utils] Déplace les classes gérants les statistiques d'exécution dans 'Profiling.h'.
