+++ arcane/src/arcane/utils/UtilsTypes.h
@@ -188,6 +188,7 @@ class MutableMemoryView;
 /*---------------------------------------------------------------------------*/
 
 class IPerformanceCounterService;
+class ForLoopTraceInfo;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
